feat(report): add rendering of table with formatted float values
Set columns with float values as 4 digits with trailing zeros and convert
all data inside dataframe into string type to prepare for rendering. Define
column alignment as centered for title and right for other columns. Use
'pdf.table' to render the table with formatted values and stylized headings.
This improves the readability and presentation of the report.