feat: update requirements.txt and add color coding to the report
Add minimum version requirements for matplotlib and pandas in requirements.txt. Also, add color coding to the detailed items and outcomes section of the report to help with clarity. Color code is red for difficult, yellow for intermediate, green for easy for difficulty; red for review, grey for low, yellow for moderate, green for high, and blue for very high for discrimination and correlation.