feat(amcreport): add option to automatically search for "Projets-QCM" directory
The commit adds an option for the user to automatically search for the "Projets-QCM" directory when creating a new file. This allows for easier navigation and selection of the desired directory. The user is prompted with a "y/n" question to determine whether they want to enable this feature. function refactored to reduce code complexity based on sonarlint suggestion.