feat(amcreport): Add functionality to retrieve OpenAI API key from environment variable
This commit adds functionality to the `amcreport.py` file to retrieve the OpenAI API key from the environment variable `OPENAI_API_KEY`. Previously, the API key was being retrieved from the `settings` dictionary, but this was not working as intended. With this change, the API key will be obtained from the environment variable if it is not already present in the `settings` dictionary.

Changes made:
- Added code to check if `openai.api_key` is None
- If `openai.api_key` is None, the code will attempt to retrieve the API key from the environment variable `OPENAI_API_KEY`
- Updated the `get_settings` function to include the retrieval of the API key from the environment variable