feat(amcreport): update 'correct' column in pd_answer
Update the 'correct' column in the pd_answer dataframe by applying a lambda function. If the value in the 'correct' column is equal to 1, set it to 1. Otherwise, set it to 0. This ensures that the 'correct' column only contains binary values. The previous behaviour would wrongly set as correct answers with '1' in the scoring part (ie. m=-1).